Extract 'questions' graphql to /graphql and /hooks folders
