+++ src/features/questions/FilterSelect.tsx
@@ -8,13 +8,13 @@ import {
   MemberGroupColors,
   MemberGroups,
 } from 'features/config/constants'
+import { useQuestionGroups } from 'features/questions/hooks/use-question-groups'
 import {
   doClaimFilter,
   doMarketFilter,
   doMemberGroupFilter,
 } from 'features/questions/utils'
 import { useNumberMemberGroups } from 'features/user/hooks/use-number-member-groups'
-import { useQuestionGroups } from 'graphql/questions/use-question-groups'
 import React from 'react'
 import { Shield, ShieldShaded } from 'react-bootstrap-icons'
 import { QuestionGroup } from 'types/generated/graphql'
